Changing all double variables to float

+++ classes/Camera.h
@@ -4,19 +4,19 @@
 class Camera {
     public: int x=0, y=0, z=0;
     public: int width=0, height=1;
-    public: double eyeX = 0;
-    public: double eyeY = 0;
-    public: double eyeZ = 5;
-    public: double centerX = 0;
-    public: double centerY = 0;
-    public: double centerZ = 0;
-    public: double upX = 0;
-    public: double upY = 1;
-    public: double upZ = 0;
-    public: double fov = 60;
-    public: double aspect = width/height;
-    public: double nearZ = 1.0;
-    public: double farZ = 20.0;
+    public: float eyeX = 0;
+    public: float eyeY = 0;
+    public: float eyeZ = 5;
+    public: float centerX = 0;
+    public: float centerY = 0;
+    public: float centerZ = 0;
+    public: float upX = 0;
+    public: float upY = 1;
+    public: float upZ = 0;
+    public: float fov = 60;
+    public: float aspect = (float)width/(float)height;
+    public: float nearZ = 1.0;
+    public: float farZ = 20.0;
     public: Camera(int width, int height);
 };
